Move the components that belong in torchchat/cli
